feat: 모달 고차컴포넌트에서 모달 배경 옵션 관리
